requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-generativeai==0.7.2
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1